+++ Scripts/generate_mature_dict.sh
@@ -1,2 +1,3 @@
+#!/bin/sh
 cd ./mature-dict/
 ./converToLua.sh ../../../totalRP3/modules/register/filter/register_mature_filter_dictionary.lua ../mature_dictionary_template.lua